refactor(#67): reflected Feedback
- PR에서의 피드백을 반영한다. 스코프와 밑의 footer에 PR 번호를 남겨놓는다.

Refs: #67
