+++ apps/web/atoms/index.ts
@@ -1,12 +1,6 @@
 export type { BlockBorderStateAtomInterface } from './blockBorderAtom';
 
 export { activedBlockGroupAtom, assembledBlockGroups, blocksStateAtom } from './blockGroupsAtom';
-export {
-  blockBorderStateAtom,
-  concurrentlyActivedSections,
-  BorderName,
-  EdgeDirectionsContstants,
-  DirectionsContstants,
-} from './blockBorderAtom';
+export { blockBorderStateAtom, concurrentlyActivedSections } from './blockBorderAtom';
 
 export { toastAtom } from './toastAtom';
