+++ apps/web/templates/layouts/sidebar/right/border-modifier/SubModifier.tsx
@@ -1,10 +1,8 @@
 import React from 'react';
 
-import { EdgeDirectionsContstants } from '@atoms/blockBorderAtom';
-
 import { useBorderMatrix, useBorderModifier } from '@hooks/index';
 
-import { DefaultHStack, DefaultVStack } from 'ui';
+import { DefaultHStack, DefaultVStack, EdgeDirectionsContstants } from 'ui';
 
 import { TemplatedColorInputWithTitlePresenter } from '../TemplatedColorInputWithTitle';
 import { TemplatedInputWithTitlePresenter } from '../TemplatedInputWithTitlePresenter';
